Fix all MSVC warnings (142 -> 0) and suppress third-party warnings
- C4305: Add f suffix to float literals (mem_queue.cpp, tutelar.cpp)
- C4267: Add static_cast<int>() for size_t->int (17 files)
- C4101: Remove unused catch variable names where safe
- C4477: Fix sprintf format strings (%ld->%zu, %lld)
- C4099: Fix struct/class forward declaration mismatches
- C4334: Use std::size_t{1} for 64-bit shift in flags.hpp
- C4834: Use std::clamp return value in fight.cpp
- C4805: Fix int==bool comparison in do_commands.cpp
- C4129: Fix unrecognized \% escape in medit.cpp
- CMake: /external:anglebrackets /external:W0, /MP, /W0 for fmt/fort
- CMake: Use yaml-cpp::yaml-cpp (deprecated target warning fix)
- Fix 5 files using "" includes for fmt -> <> for external suppression

+++ src/engine/scripting/dg_domination_helper.cpp
@@ -162,8 +162,8 @@ void process_arena_round(Trigger *trig, char *cmd)
 
 	// загрузка мобов по комнатам
 	const int total_mob_counter = dd.mob_counter_per_round.at(dd.current_round - 1);
-	const int amount_mob_each_type = total_mob_counter / var_name_mob_list.size();
-	const int amount_mob_random = total_mob_counter % var_name_mob_list.size();
+	const int amount_mob_each_type = total_mob_counter / static_cast<int>(var_name_mob_list.size());
+	const int amount_mob_random = total_mob_counter % static_cast<int>(var_name_mob_list.size());
 	if (dd.debug_mode) {
 		snprintf(buf2, kMaxStringLength, "Общее количество загружаемых мобов: %d, каждого вида: %d(x%zu), случайных мобов: %d",
 				 total_mob_counter, amount_mob_each_type, var_name_mob_list.size(), amount_mob_random);
@@ -178,7 +178,7 @@ void process_arena_round(Trigger *trig, char *cmd)
 			if (vnumum_list.empty() || mob_vnum_list.empty()) {
 				continue;
 			}
-			const auto random_room_index = number(0, vnumum_list.size() - 1);
+			const auto random_room_index = number(0, static_cast<int>(vnumum_list.size()) - 1);
 			const auto mob_index = dd.current_round - 1;
 
 			const MobVnum mob_vn = mob_vnum_list.at(mob_index);
@@ -190,13 +190,13 @@ void process_arena_round(Trigger *trig, char *cmd)
 
 	// загрузка случайных мобов в количестве amount_mob_random
 	for (auto i = 0; i < amount_mob_random; i++) {
-		const int random_type_index = number(0, dd.mob_list_to_load.size() - 1);
+		const int random_type_index = number(0, static_cast<int>(dd.mob_list_to_load.size()) - 1);
 		const auto &mob_vnum_list = dd.mob_list_to_load[random_type_index].first;
 		const auto &vnumum_list = dd.mob_list_to_load[random_type_index].second;
 		if (vnumum_list.empty() || mob_vnum_list.empty()) {
 			continue;
 		}
-		const auto random_room_index = number(0, vnumum_list.size() - 1);
+		const auto random_room_index = number(0, static_cast<int>(vnumum_list.size()) - 1);
 		const auto mob_index = dd.current_round - 1;
 
 		const MobVnum mob_vn = mob_vnum_list.at(mob_index);
